v37 fix news upload

+++ src/app/pages/admin/news/articles-list/articles-list.component.ts
@@ -8,12 +8,16 @@ import { NewsService } from '../news.service';
 import { SnackbarService } from '../../../../services/snackbar.service';
 import { ConfirmService } from '../../../../shared/confirm/confirm.service';
 import { FirebaseError } from '@angular/fire/app';
+import { MatButtonModule } from '@angular/material/button';
+import { OImage } from '../../../../shared/models/o_image.model';
+import { WarningService } from '../../../../shared/warning/warning.service';
 
 @Component({
     selector: 'app-articles-list',
     imports: [
         DatePipe,
-        MatIconModule
+        MatIconModule,
+        MatButtonModule
     ],
     templateUrl: './articles-list.component.html',
     styleUrl: './articles-list.component.scss'
@@ -24,7 +28,8 @@ export class ArticlesListComponent implements OnInit {
     newsStore = inject(NewsStore)
     newsService = inject(NewsService);
     sb = inject(SnackbarService)
-    ConfirmService = inject(ConfirmService)
+    confirmService = inject(ConfirmService)
+    warningService = inject(WarningService)
 
 
     ngOnInit(): void {
@@ -36,37 +41,36 @@ export class ArticlesListComponent implements OnInit {
             })
     }
     onEdit(articleId: string) {
-        // console.log(articleId)
-        this.fs.getDoc(`articles/${articleId}`)
-            .subscribe((article: Article) => {
-                this.newsStore.setArticle(article);
-                this.newsService.articleActivated.next(article)
-            })
+        this.newsService.getArticle(articleId)
+        // this.fs.getDoc(`articles/${articleId}`)
+        //     .subscribe((article: Article) => {
+        //         this.newsStore.setArticle(article);
+        //         this.newsService.articleChanged.next(article)
+        //     })
     }
 
-    onDelete(articleId: string, articleTitle: string) {
-        this.ConfirmService.getConfirmation(articleTitle)
-            .then((res: boolean) => {
-                if (res) {
-                    this.fs.deleteDoc(`articles/${articleId}`)
-                        .then((res: any) => {
-                            this.sb.openSnackbar(`article deleted`)
-                        })
-                        .catch((err: FirebaseError) => {
-                            console.log(err)
-                            this.sb.openSnackbar(`operation failed due to: ${err.message}`)
-                        })
-                }
-            })
-        // console.log(articleId)
-        // const dialogRef = this.dialog.open(ConfirmComponent, {
-        //     data: {
-        //         doomedElement: articleId
-        //     }
-        // })
-        // dialogRef.afterClosed().subscribe((res: boolean) => {
-        //     if (res) {
-        //     }
-        // })
+    onDelete(articleId: string, articleTitle: string, oImage: OImage) {
+        if (oImage) {
+            this.warningService.showWarning('remove image before removing news item')
+
+        } else {
+            console.log('proceed')
+            this.confirmService.getConfirmation(articleTitle)
+                .then((res: boolean) => {
+                    if (res) {
+                        this.fs.deleteDoc(`articles/${articleId}`)
+                            .then((res: any) => {
+                                this.sb.openSnackbar(`article deleted`)
+                            })
+                            .catch((err: FirebaseError) => {
+                                // console.log(err)
+                                this.sb.openSnackbar(`operation failed due to: ${err.message}`)
+                            })
+                    }
+                })
+        }
+    }
+    onAddNewArticle() {
+        this.newsService.articleChanged.emit(null);
     }
 }
